requirements.txt
<<<<<<< HEAD
Flask==2.3.2
Pillow==10.0.0
transparent-background==1.3.2
Werkzeug==2.3.6
pillow-heif==0.10.0
=======
Flask
Pillow
transparent-background
Werkzeug
pillow-heif
>>>>>>> c443152d9574b52dcf02d4094b813274dbc16359
